remove flag defaults
if we send in zeros then the fn server will correctly pick defaults. this is
preferable anyway, in case fn defaults change. so this works the same as
before in practice for route update/create.

the func file now no longer gets flags that previously had defaults written to
it